backend/requirements.txt
# DreamLand Backend Dependencies
# Core FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# OpenAI integration
openai==1.3.5

# Async and background tasks
aiofiles==23.2.1
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# CORS and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional: Redis for production async tasks
# redis==5.0.1
# rq==1.15.1

# Development
pytest==7.4.3
httpx==0.25.1